src/session_service/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.31
Flask-Cors==4.0.1
flask_apscheduler
redis
psycopg2-binary